ARM: (Experimental) Provide Estimated CPU Capacity measure
Bsed upon the CPU Power of a core, computes a capacity measure
between 0 and 1024 scaling in line with the frequency using a
simple linear scale derived from the maximum frequency reported
by CPUFreq.

Scaling CPU Power with frequency and estimated capacity gives an
estimate of the amount of potential compute capacity available to
a specific core relative to any other in the system.

Change-Id: I8048a23fe5999536b6325a5ec64549dd69f5a865
backported for Mako Hybrid kernel from LKML

Signed-off-by: Paul Reioux <reioux@gmail.com>